feat(core): migrate analytics package to junit5
