+++ core/src/test/kotlin/com/malinskiy/marathon/analytics/metrics/MetricsProviderFactoryTest.kt
@@ -2,7 +2,6 @@ package com.malinskiy.marathon.analytics.metrics
 
 import com.malinskiy.marathon.analytics.external.MetricsProviderFactory
 import com.malinskiy.marathon.analytics.external.NoOpMetricsProvider
-import com.malinskiy.marathon.analytics.external.influx.InfluxMetricsProvider
 import com.malinskiy.marathon.device.DeviceFeature
 import com.malinskiy.marathon.device.DeviceProvider
 import com.malinskiy.marathon.execution.AnalyticsConfiguration
@@ -97,50 +96,5 @@ class MetricsProviderFactoryTest {
         val factory = MetricsProviderFactory(configuration)
         val metricsProvider = factory.create()
         metricsProvider shouldBeInstanceOf NoOpMetricsProvider::class
-    }
-
-    @Test
-    fun shouldReturnInfluxProvider() {
-        val analyticsConfiguration = AnalyticsConfiguration.InfluxDbConfiguration(
-            "http://influxdb.local",
-            "user",
-            "password",
-            "db",
-            AnalyticsConfiguration.InfluxDbConfiguration.RetentionPolicyConfiguration.default
-        )
-        val configuration = Configuration(
-            name = "",
-            outputDir = File(""),
-            analyticsConfiguration = analyticsConfiguration,
-            poolingStrategy = null,
-            shardingStrategy = null,
-            sortingStrategy = null,
-            batchingStrategy = null,
-            flakinessStrategy = null,
-            retryStrategy = null,
-            filteringConfiguration = null,
-            ignoreFailures = null,
-            isCodeCoverageEnabled = null,
-            fallbackToScreenshots = null,
-            strictMode = null,
-            uncompletedTestRetryQuota = null,
-            testClassRegexes = null,
-            includeSerialRegexes = null,
-            excludeSerialRegexes = null,
-            testBatchTimeoutMillis = null,
-            testOutputTimeoutMillis = null,
-            debug = null,
-            screenRecordingPolicy = null,
-            vendorConfiguration = object : VendorConfiguration {
-                override fun testParser(): TestParser? = null
-                override fun deviceProvider(): DeviceProvider? = null
-                override fun logConfigurator(): MarathonLogConfigurator? = null
-                override fun preferableRecorderType(): DeviceFeature? = null
-            },
-            analyticsTracking = false
-        )
-        val factory = MetricsProviderFactory(configuration)
-        val metricsProvider = factory.create()
-        metricsProvider shouldBeInstanceOf InfluxMetricsProvider::class
     }
 }
